# Replace Fnv hash with hashbrown