--- a/Cargo.toml
+++ b/Cargo.toml
@@ -24,7 +24,7 @@ include = [
 travis-ci = { repository = "greyblake/whatlang-rs", branch = "master" }
 
 [dependencies]
-fnv = "1.0.6"
+hashbrown = "0.1"
 
 [dev-dependencies]
 serde_json = "1.0.32"